Objective-C interoperability and Vector Resources.
