+++ PKHUD/Images.xcassets/checkmark.imageset/Contents.json
@@ -2,13 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "scale" : "1x",
-      "filename" : "checkmark.png"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x",
-      "filename" : "checkmark@2x.png"
+      "filename" : "checkmark.pdf"
     }
   ],
   "info" : {
