+++ PKHUD/Images.xcassets/cross.imageset/Contents.json
@@ -2,13 +2,7 @@
   "images" : [
     {
       "idiom" : "universal",
-      "scale" : "1x",
-      "filename" : "cross.png"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x",
-      "filename" : "cross@2x.png"
+      "filename" : "cross.pdf"
     }
   ],
   "info" : {
